refactor(apps/client): HASHI-165 bearer token parser 이동
