+++ apps/client/src/features/auth/api/reissueToken.ts
@@ -1,7 +1,7 @@
 import { ApiError, HttpStatusError } from '@/shared/api/apiError'
 import { apiClient } from '@/shared/api/apiClient'
 import { isErrorResponse, isSuccessResponse } from '@/shared/api/types'
-import { extractBearerToken } from '@/features/auth/utils/extractBearerToken'
+import { extractBearerToken } from '@/shared/api/extractBearerToken'
 
 export interface TokenReissueResult {
   accessToken: string
