+++ apps/client/src/shared/api/extractBearerToken.ts
